Added User validation if user is user he can edit his data

+++ src/main/java/nl/novi/gamenight/Services/UserService.java
@@ -12,7 +12,9 @@ import nl.novi.gamenight.exceptions.UserNotUniqueException;
 import nl.novi.gamenight.security.MyUserDetails;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
@@ -82,6 +84,10 @@ public class UserService {
 
     /*TODO Admin And owning user*/
     public ResponseEntity deleteUserByID(Long id) {
+        if (!checkOwningUser(id)) {
+            return ResponseEntity.badRequest().body("User is not authorized");
+        }
+
         Optional<User> ifExist = userRepository.findById(id);
         if (ifExist.isPresent()) {
             userRepository.deleteById(id);
@@ -94,7 +100,9 @@ public class UserService {
     /*TODO Admin And owning user*/
     public ResponseEntity<Object> updateUserNameByID(@Validated Long id, UserInputDto updatedUser, BindingResult bindingResult) {
         Optional<User> ifExist = userRepository.findById(id);
-
+        if (!checkOwningUser(id)) {
+            return ResponseEntity.badRequest().body("User is not authorized");
+        }
         if (ifExist.isPresent()) {
             if (bindingResult.hasErrors()) {
                 Map<String, String> errors = new HashMap<>();
@@ -127,6 +135,24 @@ public class UserService {
         user.setUsername(userInputDto.username);
         user.setPassword(encoder.encode(userInputDto.password));
         return user;
+    }
+
+    private boolean checkOwningUser(Long userID) {
+        var user = userRepository.findById(userID);
+        if (user.isEmpty()) {
+            return false;
+        }
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        MyUserDetails userDetails = (MyUserDetails) principal;
+
+        if(userID.equals(userDetails.getUserId())) {
+            return true;
+        }
+        Collection<? extends GrantedAuthority> authorities = userDetails.getAuthorities();
+
+        return authorities.stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ADMIN"));
     }
 }
